scheduler: Take the launch time as reference time
This will ensure that no events are ignored when we are called again an
hour later. The previous implementation was sensitive to delays of other
preceding tasks, which resulted in them not being called every hour, but
sometimes every second hour. This is bad for health checks.
This implementation will be much more deterministic.