timer: Simplify systemd timer
We want to be called some time after boot (the previous Persistent=true
setting made the system call us immediately after the machine was off
for an hour), and then every hour.
With these settings, the timer does not need to be persistent anymore,
unless the system in question never runs for longer than 5min, which
seems extremely unlikely.